Create check_if_server_is_up_v3.sh
One potential way to improve the efficiency of this script is to use the -n option with the ping command to send only a single packet instead of five. This will decrease the time required to perform the check, especially if the server is unreachable. Additionally, instead of using the ping command, you can use the nc (netcat) command to check if a specific port is open on the server. This can be faster than using ping since it doesn't require the same amount of time to wait for a response.